Optimize the Xoroshiro64Star subroutine
This saves 12 bytes and 846 cycles (down from 4031 to 3185 cycles for
the initial test state), at the cost of making the code trickier and
having it take a variable amount of time depending on the current state
(from 2691 for all bits 0, to 3907 for all bits 1).

Overview:
- use immediate values for the fixed factor inside the loop
- use a sentinel bit instead of counting with Y
- keep the high byte of the temp result word in A
- reorder some things, and some other minor changes

This also fixes the comment character used in the printSize macro.