Speed up the test output code (using a larger buffer)
Previously, most of the tests had to wait for the next frame before
showing any results, since the buffer was only a single tile long.

Now, the buffer can hold up to 2 whole lines, which seems to be enough
that none of the tests end up waiting for buffer space anymore.

Thus, the tests finish much faster, and the output more clearly shows
just how slow the jump functions are, as compared to the others.