Deduplicate the test code
This removes a lot of duplicated routines, and uses nicer names for the
generic versions that remain. This requires a bit more initialization
for each test, but a couple of macros avoids most of that repetition.

This ends up saving 101 bytes of code, at the expense of 6 bytes in ZP.